Add logic tracking

+++ transaction/models.py
@@ -51,7 +51,7 @@ class TruckDirection(models.Model):
                               on_delete=models.CASCADE, related_name='truck_direction')
     place = models.ForeignKey(Place, default=None,
                               on_delete=models.CASCADE, related_name='place_truck_direction')
-    takes_time = models.IntegerField()
+    takes_time = models.DecimalField(decimal_places=2, max_digits=12)
     amount_km = models.DecimalField(decimal_places=2, max_digits=12)
 
     created_at = models.DateTimeField(
